requirements.txt
openai
python-dotenv
langchain
langchain_openai
langchain_community
langchain_text_splitters
requests
beautifulsoup4
FastAPI
redis
chromadb
faiss-cpu